Add login endpoint and integrate security

+++ src/main/java/com/thangle/domain/book/BookService.java
@@ -1,11 +1,13 @@
 package com.thangle.domain.book;
 
+import com.thangle.domain.auth.AuthsProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import com.thangle.persistence.book.BookStore;
 
 import static com.thangle.domain.book.BookError.supplyBookNotFound;
+import static com.thangle.error.CommonError.supplyForbiddenError;
 import static com.thangle.domain.book.BookValidation.validateBook;
 
 import java.time.Instant;
@@ -17,6 +19,7 @@ import java.util.UUID;
 public class BookService {
 
     private final BookStore bookStore;
+    private final AuthsProvider authsProvider;
 
     public List<Book> findAll() {
         return bookStore.findAll();
@@ -32,6 +35,7 @@ public class BookService {
 
     public Book create(final Book book) {
         validateBook(book);
+        book.setUserId(authsProvider.getCurrentUserId());
         book.setCreatedAt(Instant.now());
         return bookStore.save(book);
     }
@@ -40,18 +44,29 @@ public class BookService {
         final Book book = findById(id);
         validateBook(updatedBook);
 
+        validateBookActionPermission(book);
+
         book.setTitle(updatedBook.getTitle());
         book.setAuthor(updatedBook.getAuthor());
         book.setDescription(updatedBook.getDescription());
         book.setUpdatedAt(Instant.now());
         book.setImage(updatedBook.getImage());
-        book.setUserId(updatedBook.getUserId());
 
         return bookStore.save(book);
     }
 
     public void deleteById(final UUID id) {
         final Book book = findById(id);
+
+        validateBookActionPermission(book);
+
         bookStore.deleteById(book.getId());
+    }
+
+    private void validateBookActionPermission(final Book book) {
+        if (authsProvider.getCurrentUserRole().equals("ROLE_CONTRIBUTOR")
+                && !authsProvider.getCurrentUserId().equals(book.getUserId())) {
+            throw supplyForbiddenError().get();
+        }
     }
 }
